TJShared: - In the pthreads implementation of CriticalSection, a thread that would call Enter() twice on the same mutex would block (on itself). The mutex created is now of a 'PTHREAD_MUTEX_RECURSIVE' type, which allows this (and also uses a lock count to match the Windows CRITICALSECTION-behaviour)
TJScript:
- Because instances of the ScriptValue<T> class are not castable from Scriptable to ScriptValue<T> across module boundaries, a new method of unboxing these values is now implemented (through a class called ScriptAny, using tj::shared::Any). This also makes the behaviour of TJScript more consistent with other Any-using parts of the software.




git-svn-id: https://svn.pixelspark.nl/svn/corespark/trunk@547 c9a305e2-9f8a-4be3-8c95-b0e5bab41dc9

+++ TJShared/src/tjtime.cpp
@@ -186,7 +186,7 @@ void Timestamp::Now() {
 }
 
 String Timestamp::ToString() const {
-	return Stringify(ToMicroSeconds());
+	return Stringify((int)ToMicroSeconds());
 }
 
 String Timestamp::ToHexString() const {
